Removed ethereum-rpc-rx module and migrated its tests to coroutines
