+++ ethereum-rpc-co/build.gradle
@@ -19,9 +19,6 @@ dependencies {
     // Bivrost
     implementation "com.github.gnosis.bivrost-kotlin:bivrost-solidity-types:$versions.bivrost"
 
-    testImplementation "org.mockito:mockito-core:$versions.mockito"
-    testImplementation "org.mockito:mockito-inline:$versions.mockito"
     testImplementation "io.mockk:mockk:1.9.3"
-    testImplementation "org.jetbrains.kotlinx:kotlinx-coroutines-test:1.3.1"
-    testImplementation project(":utils-testing")
+    testImplementation "org.jetbrains.kotlinx:kotlinx-coroutines-test:1.3.5"
 }
