+++ ethereum-rpc-retrofit/src/main/java/pm/gnosis/ethereum/rpc/retrofit/RetrofitEthereumRpcApi.kt
@@ -8,17 +8,17 @@ import retrofit2.http.POST
 
 interface RetrofitEthereumRpcApi {
     @POST(".")
-    fun receipt(@Body jsonRpcRequest: JsonRpcRequest): Response<JsonRpcTransactionReceiptResult>
+    suspend fun receipt(@Body jsonRpcRequest: JsonRpcRequest): JsonRpcTransactionReceiptResult
 
     @POST(".")
-    fun block(@Body jsonRpcRequest: JsonRpcRequest): Response<JsonRpcBlockResult>
+    suspend fun block(@Body jsonRpcRequest: JsonRpcRequest): JsonRpcBlockResult
 
     @POST(".")
-    fun transaction(@Body jsonRpcRequest: JsonRpcRequest): Response<JsonRpcTransactionResult>
+    suspend fun transaction(@Body jsonRpcRequest: JsonRpcRequest): JsonRpcTransactionResult
 
     @POST(".")
-    fun post(@Body jsonRpcRequest: JsonRpcRequest): Response<JsonRpcResult>
+    suspend fun post(@Body jsonRpcRequest: JsonRpcRequest): JsonRpcResult
 
     @POST(".")
-    fun post(@Body jsonRpcRequest: Collection<JsonRpcRequest>): Response<Collection<JsonRpcResult>>
+    suspend fun post(@Body jsonRpcRequest: Collection<JsonRpcRequest>): Collection<JsonRpcResult>
 }
