--- ethereum-rpc-rx/build.gradle
@@ -1,31 +1,0 @@
-apply plugin: 'com.android.library'
-apply plugin: 'kotlin-android'
-apply plugin: 'kotlin-kapt'
-apply from: '../buildsystem/coverageReport.gradle'
-
-dependencies {
-    // We specify junit before anything else to make sure that this version is prioritized over the
-    // one bundled in the android.jar
-    testImplementation "junit:junit:$versions.junit"
-
-    implementation project(":ethereum")
-    implementation project(":ethereum-rpc")
-    implementation project(":models")
-    implementation project(":utils")
-
-    //Rx
-    implementation "io.reactivex.rxjava2:rxjava:$versions.rxjava"
-
-    // Moshi
-    implementation "com.squareup.moshi:moshi:$versions.moshi"
-    kapt "com.squareup.moshi:moshi-kotlin-codegen:$versions.moshi"
-
-    // Bivrost
-    implementation "com.github.gnosis.bivrost-kotlin:bivrost-solidity-types:$versions.bivrost"
-
-    testImplementation "org.mockito:mockito-core:$versions.mockito"
-    testImplementation "org.mockito:mockito-inline:$versions.mockito"
-    testImplementation "io.mockk:mockk:1.9.3"
-    testImplementation "org.jetbrains.kotlinx:kotlinx-coroutines-test:1.3.1"
-    testImplementation project(":utils-testing")
-}
