--- ethereum-rpc-rx/src/main/AndroidManifest.xml
@@ -1,1 +1,0 @@
-<manifest package="pm.gnosis.ethereum.rpc.rx" />
