+++ settings.gradle
@@ -17,5 +17,4 @@ include ':security'
 include ':utils'
 include ':utils-ethereum'
 include ':utils-testing'
-include ':ethereum-rpc-rx'
 include ':ethereum-rpc-co'
